Do not single-step within syscall
Using PTRACE_SINGLESTEP on a syscall-enter-stop can have unexpected
effects. On x86_64, it seems to skip over the syscall-exit-stop, while
on aarch64, it seems to cause a spurious SIGTRAP with empty si_code and
without si_addr.

Avoid these issues by using PTRACE_SYSCALL instead of PTRACE_SINGLESTEP
when the syscall-stop is not a syscall-exit-stop.